+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # osgl-csv CHANGE LOG
 
 1.0.0 - Initial Version
+* depend on osgl-tool-1.18.0
 * Read csv file into Map/AdaptiveBean/POJO list