+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # osgl-excel CHANGE LOG
 
-1.5.0
+1.5.0 - 30/Oct/2018
 * Add osgl.ext.list file for IO integration
 * Support write big data to Excel file #15
 * Support specify cell data format by header name #14